requirements.txt
fastapi
uvicorn
openai
python-dotenv
pymupdf
python-multipart
